LayoutTests/js/constant-count-expected.txt

This test checks exceptional cases for constant counting in the parser.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'a' is all good
PASS f() is 10
PASS successfullyParsed is true

TEST COMPLETE
